Create persistence_layer for models
