+++ app/models/experience.rb
@@ -1,15 +1,2 @@
-class Experience
-  include Neo4j::ActiveNode
-
-  has_one :in, :host, model_class: false
-
-  has_many :out, :times, model_class: ExperienceTime
-
-  property :id
-  property :name
-  property :location
-  property :description
-
-  validates :name,        presence: true, length: { maximum: 50 }
-  validates :location,    presence: true, length: { maximum: 30 }
+class Experience < PersistenceLayers::Neo4j::Experience
 end
